feat!: rewrite and add support for local files via http/https
This is a full rewrite of Rustfoil which includes upgrading/replacing dependencies with newer versions or stable alternatives.

This also includes splitting the CLI into a lib and the CLI itself which makes it easier to programmatically make use of the logic of Rustfoil.

I also added support for indexing local files which are served by an http server, this is useful if you want to host files locally but still have an index for the motd or requried themes etc.